Added Visual Studio 2013 project. Moved projections out of transforms.h

+++ src/mathematics.c
@@ -22,7 +22,7 @@
 #include <math.h>
 #include "mathematics.h"
 
-float uniformi( void )
+int uniformi( void )
 {
 	return rand( );
 }
